% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonGPBayes.R
\name{as.poisGPBayes0.poisGPBayes}
\alias{as.poisGPBayes0.poisGPBayes}
\title{Coercion Method}
\usage{
\method{as.poisGPBayes0}{poisGPBayes}(object, ...)
}
\arguments{
\item{object}{An object with class \code{"poisGPBayes0"}}

\item{...}{Further arguments to be passed to \code{poisGPBayes0}.}
}
\value{
An object with class \code{"poisGPBayes0"}.
}
\description{
Coercion method.
}
\note{
Inasmuch the classes \code{"poisGPBayes"} and
\code{"poisGPBayes0"} contain named list with consistent naming
rules this method nearly only changes the class of the object. It
could be made to work in this way in the future.
}
