% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFuns.R
\name{as.poisGPBayes0}
\alias{as.poisGPBayes0}
\title{Compute Return Levels using an Extreme-Value Model}
\usage{
as.poisGPBayes0(object, ...)
}
\arguments{
\item{object}{The object to be coerced.}

\item{...}{Not used yet.}
}
\value{
An object containing the return levels such as an object
inheriting from \code{data.frame}.

This method is typically used with a \code{poisGPBayes} object to
access the method written for Poor Man's objects.

An object with class \code{"poisGPBayes0"} inheriting from
\code{"Bayes0"}.
}
\description{
Compute Return Levels and confidence or credible intervals on these.
}
