% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRL.R
\name{autolayer.predRL}
\alias{autolayer.predRL}
\title{\code{autolayer} Method for Predictive Return Levels}
\usage{
\method{autolayer}{predRL}(object, aes = FALSE, ...)
}
\arguments{
\item{object}{An object with class \code{"predRL"} corresponding
to predictive Return Levels.}

\item{aes}{Logical. If \code{TRUE} the colour and linetype of the
line are added to the aesthetic in odred to appear in
legends.}

\item{\dots}{Further arguments to be passed to \code{geom_line}.}
}
\value{
An ggplot layer
}
\description{
Predictive Return Level curve
}
\details{
The argument \code{object} contains a table of computed predictive
Return Levels, in correspondence with an exceedance
probability. The layer shows the tail-quantile function for the
predictive distribution.
}
\note{
In practice the predictive distribution for an Extreme-Value
model has usually an heavy tail. Therefore at least for small
probabilities of exceedance \eqn{p} the curve is convex (upward
concave).
}
\seealso{
\code{\link{autoplot.predRL}}.
}
