% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotpoisGP.R
\name{autoplot.poisGPBayes0}
\alias{autoplot.poisGPBayes0}
\title{\code{autoplot} Method for the \code{"poisGPBayes"} S3
Class}
\usage{
\method{autoplot}{poisGPBayes0}(object,
         which = c("RL", "pp"),
         level = 0.70,
         points = c("H", "p", "none"), a = 0.5,
         allPoints = FALSE,
         trace = 0, ...)
}
\arguments{
\item{object}{An object with class \code{"poisGPBayes"} describing
inference results for a Poisson-GP model.}

\item{which}{The type of plot wanted.}

\item{level}{Credible level(s).}

\item{points, a, allPoints}{\code{\link[potomax]{autoplot.poisGP}}.}

\item{trace}{Integer level of verbosity.}

\item{...}{Further arguments passed to \code{RL}.}
}
\value{
An object with class \code{"gg"} inheriting from
\code{"ggplot"} which is typically used through the \code{print}
method.
}
\description{
Autoplot an object with call \code{"poisGPBayes"}.
}
\examples{
## ========================================================================
## Use the Garonne data from Renext
## ========================================================================
fitBayes <- poisGPBayes(data = Garonne$OTdata$Flow,
                        threshold = 2500,
                        effDuration = Garonne$OTinfo$effDuration)

## ========================================================================
## S3 methods: Return Levels, ...
## ========================================================================
RLBayes <- RL(fitBayes)
autoplot(fitBayes)            ## nearly the same plot
}
