% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRL.R
\name{autoplot.predRL}
\alias{autoplot.predRL}
\title{\code{autoplot} Method for Predictive Return Levels}
\usage{
\method{autoplot}{predRL}(object, points = c("none", "p"), a = 0.5, ...)
}
\arguments{
\item{object}{An object with class \code{"predRL"} corresponding
to predictive Return Levels.}

\item{points}{Character indicating if the data attached to the
object (if any) should be shown (value \code{"p"}) or not (value
\code{"none"}).}

\item{a}{The value of the argument \code{a} of the function
\code{\link[stats]{ppoints}} when computing the plotting postiions.}

\item{...}{Not used yet.}
}
\value{
An object with class \code{"gg"} inheriting from
\code{"ggplot"}.
}
\description{
Predictive Return Level plot using an object containing the Return
Levels.
}
\details{
The argument \code{object} contains a table of computed predictive
Return Levels, in correspondence with an exceedance
probability. The plot shows the tail-quantile function for the
predictive distribution.
}
\note{
In practice the predictive distribution for an Extreme-Value
model has usually an heavy tail. Therefore at least for small
probabilities of exceedance \eqn{p} the curve is convex (upward
concave).
}
\seealso{
\code{\link{autoplot.predRLList}} to compare predictive
plots: for several predicted durations, several priors, several
datasets, ...
}
