% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotGEV.R
\name{autoplot.RL.GEVBayes}
\alias{autoplot.RL.GEVBayes}
\title{\code{autoplot} Method for the \code{"RL.GEVBayes"} S3
Class}
\usage{
\method{autoplot}{RL.GEVBayes}(object, level = 0.70,
        xVar = c("T", "p"),
        trace = 0,... )
}
\arguments{
\item{object}{An object with class \code{"RL.GEVBayes"} describing
Return Levels and related Bayesian inference results for a GEV
model. This object is generated by \code{\link{RL.GEVBayes0}}.}

\item{level}{Credible level.}

\item{xVar}{The variable used to match the \eqn{x}-axis. The value
\code{"T"} will lead to use the column named \code{"Period"} of
\code{object} while the value \code{"p"} will lead to use the
column named \code{"Prob"}.}

\item{trace}{Integer level of verbosity.}

\item{...}{Further arguments passed to \code{RL}.}
}
\value{
An object with class \code{"gg"} inheriting from
\code{"ggplot"}, typically used through the \code{print} method.
}
\description{
Autoplot an object with class \code{"RL.GEVBayes"} containing
Return Levels and related inference results.
}
\seealso{
\code{\link{autoplot.GEVBayes0}}.
}
