% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotpoisGP.R
\name{autoplot.RL.poisGPBayes}
\alias{autoplot.RL.poisGPBayes}
\title{\code{autoplot} Method for the \code{"RL.poisGPBayes"} S3
Class}
\usage{
\method{autoplot}{RL.poisGPBayes}(object, ...)
}
\arguments{
\item{object}{An object with class \code{"RL.poisGPBayes"}
describing Return Levels and related Bayesian inference results
for a Poisson-GP model.}

\item{...}{Further arguments passed to \code{RL}.}
}
\value{
An object with class \code{"gg"} inheriting from
\code{"ggplot"}, typically used through the \code{print} method.
}
\description{
Autoplot an object with class \code{"RL.poisGPBayes"} containing
Return Levels and related inference results.
}
\seealso{
\code{\link{autoplot.poisGPBayes0}}.
}
