% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGPD.R
\name{CGPD}
\alias{CGPD}
\alias{dCGPD}
\alias{pCGPD}
\alias{qCGPD}
\alias{rCGPD}
\title{Density, Distribution Function, Quantile Function and
Random Generation for the Five-Parameter Compound Generalized
Pareto Distribution (CGPD)}
\usage{
dCGPD(x, loc = 0.0, scale = 1.0, shape = 0.0,
      scaleN, shapeN, EN, IDN, log = FALSE)

pCGPD(q, loc = 0.0, scale = 1.0, shape = 0.0,
      scaleN, shapeN, EN, IDN, lower.tail = TRUE)

qCGPD(p, loc = 0.0, scale = 1.0, shape = 0.0,
      scaleN, shapeN, EN, IDN, lower.tail = TRUE)

rCGPD(n, loc = 0.0, scale = 1.0, shape = 0.0,
      scaleN, shapeN, EN, IDN)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{loc}{Location parameter. Numeric vector of length one.}

\item{scale}{Scale parameter. Numeric vector of length one.}

\item{shape}{Shape parameter. Numeric vector of length one.}

\item{scaleN}{Scale of the GPD for the \eqn{N} part. Along with
\code{shapeN} it provides the parameterisation for the
Binomial-Poisson-Negative Binomial familly.}

\item{shapeN}{Shape of the GPD for the \eqn{N} part. Along with
\code{scaleN} it provides the parameterisation for the
Binomial-Poisson-Negative Binomial familly.}

\item{EN}{Expectation of \eqn{N}. Along with \code{IDN} it
provides an alternative parameterisation for the \eqn{N} part.}

\item{IDN}{Index of Dispersion of \eqn{N}. Along with \code{EN} it
provides an alternative parameterisation for the \eqn{N} part.}

\item{log}{Logical; if \code{TRUE}, densities \code{p} are
returned as \code{log(p)}.}

\item{p}{Vector of probabilities.}

\item{n}{Sample size.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are
P[X <= x], otherwise, P[X > x].}
}
\value{
A numeric vector with length equal to the length of the
first argument or of the parameters.
}
\description{
Density, distribution function, quantile function and random
generation for the five-parameter Compound Generalized Pareto
Distibution (CGPD).
}
\details{
This distribution is that of the maximum \eqn{M} of
\eqn{N} i.i.d. r.vs \eqn{X_i} with distribution
\eqn{\textrm{GPD}(\mu,\,\sigma,\,\xi)}{GPD(\mu, \sigma, \xi)}
where \eqn{N} is a r.v. with non-negative integer values,
independent of the sequence \eqn{X_i}, and having a
\emph{Binomial}, \emph{Poisson} or \emph{Negative Binomial}
distribution. The distribution of \eqn{N} can be parameterized by
using two parameters \eqn{\mu_N}{\mu_N} and
\eqn{\sigma_N}{\sigma_N} in a GPD style, or alternatively by using
the two parameters \eqn{\mathrm{E}(N)}{EN} and
\eqn{\textrm{ID}(N)}{IDN} representing the expectation and the
index of dispersion of \eqn{N}. The three cases \emph{Binomial},
\emph{Poisson} and \emph{Negative Binomial} correspond to
\eqn{\textrm{ID}_N < 0}{IDN < 0}, \eqn{\textrm{ID}_N = 0}{IDN = 0}
and \eqn{\textrm{ID}_N > 0}{IDN > 0}.
}
\section{Caution}{
 This distribution is of \emph{mixed-type}. It
has a probability mass at \eqn{-\infty}{-Inf} corresponding to the
possibility that \eqn{N=0} in which case \eqn{M} is the maximum of
an empty set, taken as \eqn{-\infty}{-Inf} corresponding to
\code{max(mumeric(0))}. Consequently a sample drawn by using
\code{rCGPD} contains \code{-Inf} values with positive
probability.
}

\examples{
set.seed(1)
ExpN <- runif(1)
IDN <- rexp(1, rate = 1)
scaleN <- 1 / ExpN
shapeN <- (IDN - 1) / ExpN
loc <- rnorm(1, mean = 0, sd = 10); scale <-  rexp(1)
shape <- rnorm(1 , mean = 0, sd = 0.1)
mass <- pCGPD(-Inf, scaleN = scaleN, shapeN = shapeN,
           loc = loc, scale = scale, shape = shape) 
q <- qCGPD(p = c(mass + 0.001, 0.999), scaleN = scaleN, shapeN = shapeN,
           loc = loc, scale = scale, shape = shape)
x <- seq(from = q[1] - 1, to = q[2], length.out = 200)
F <- pCGPD(x, scaleN = scaleN, shapeN = shapeN,
           loc = loc, scale = scale, shape = shape)
plot(x, F, type = "l", xlab = "", ylab = "", ylim = c(0, 1),
     col = "orangered")
abline(h = mass, col = "red")
f <- dCGPD(x, scaleN = scaleN, shapeN = shapeN,
           loc = loc, scale = scale, shape = shape)
plot(x, f, type = "l", col = "SteelBlue3", xlab = "", ylab = "")
title(main = sprintf(paste("ExpN = \%4.1f IDN = \%4.2f,",
                           "loc = \%4.1f, scale = \%4.2f, shape = \%4.2f"),
                     ExpN, IDN, loc, scale, shape))

}
\references{
Yves Deville (2019)  "Bayesian Return Levels in Extreme-Value Analysis"
\emph{IRSN technical report}.
}
\author{
Yves Deville
}
