% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parNames.R
\name{checkParNames}
\alias{checkParNames}
\title{Check a Candidate Vector or Parameter Names for an
Extreme-Value Model}
\usage{
checkParNames(parNames, model, all = TRUE, trace = 0)
}
\arguments{
\item{parNames}{Candidate vector of parameter names.}

\item{model}{A possible Extreme-Value model. This is
case-insensitive.}

\item{all}{Logical. If \code{TRUE}, the vector is assumed to
contain all the parameter names, possibly in a different order.}

\item{trace}{Integer level of verbosity.}
}
\value{
A list with the following items

\item{rule}{

The name of the identified rule for naming the parameters. For
instance, \code{"names"} corresponds to the rule used in this
package, \code{"greek"} corresponds using the names of the greek
letters associated to the parameters.

}
\item{inParNames}{

A copy of the vector \code{parNames} that was given on input.

}
\item{parNames}{

The vector of names corresponding to the rule \code{"names"} hence
the convention used in this package.

}
\item{indIn, indOut}{

Vectors of integer indices relating to the conventional vector of
names. \code{indIn} corresponds to the given names, possibly
translated into the conventional rule. \code{indOut} corresponds
to the omitted names, and can only have a positive length when
\code{all} is \code{FALSE} on input.

}
}
\description{
Given a candidate vector of parameter names, check that these
match a known naming rule suitable for the given model.
}
\examples{
checkParNames(parNames = c("sigma", "mu", "xi"), model = "GEV")
checkParNames(parNames = c("sigma", "xi"), model = "GEV", all = FALSE)
checkParNames(parNames = c("lambda", "sigma", "xi"), model = "poisgp")
}
