% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscTools.R
\name{credInt}
\alias{credInt}
\title{Find a Credible Interval for a Given Unimodal Probability
Distribution}
\usage{
credInt(dist, level = 0.95,
        type = c("HPD", "eqtail", "both"),
        plot = FALSE, ...)
}
\arguments{
\item{dist}{Either a character describing a distribution such as
\code{"norm"} or \code{"gamma"}, or an object with class
\code{"density"} or a numeric vector assumed to provide a sample
of the distribution.}

\item{level}{Numeric value specifying the probability level.  A
vector of length \eqn{> 1} is not accepted yet.}

\item{type}{When set to \code{"HPD"}, the credible interval is
Highest Posterior Density interval. When set to \code{"eqtail"}
the bounds are found by using the cumulative distribution function
and by allocating the same probability to the left and right
tails.}

\item{plot}{Logical. If \code{TRUE} a plot will be shown.}

\item{...}{Other arguments to be passed to the distribution
functions: density, quantile.}
}
\value{
A numeric vector with the two bounds or an array when
\code{type} is \code{"both"}.
}
\description{
Credible interval for a unimodal density, usually a posterior
distribution in a Bayesian framework.
}
\details{
In some applications, \code{dist} will be a specific distribution
within a parametric family. This can occur for instance when a
conjugacy property holds.
}
\note{
When \code{dist} is intended to describe a distribution
within a parametric family, it must be taken as the character
string which is prefixed by \code{"d"}, \code{"p"} and \code{"q"}
to get the density, the distribution and the quantile
functions. The values of the parameters can passed through the
dots \dots See \bold{Examples}.
}
\examples{
credInt("gumbel", level = 0.95, plot = TRUE)
credInt("gamma", level = 0.95, plot = TRUE, shape = 4)
credInt("gumbel", level = 0.95, plot = TRUE, type = "both")
x <- rgamma(n = 6000, shape = 4)
credInt(x, level = 0.95, plot = TRUE)
## nearly the same
d <- density(x)
credInt(x, level = 0.95, plot = TRUE)
}
