% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscTools.R
\name{formatLevel}
\alias{formatLevel}
\title{Format Levels of Confidence or Probabilities}
\usage{
formatLevel(level)
}
\arguments{
\item{level}{Numeric vector of levels of confidence or
probabilities between \code{0} and \code{1}.}
}
\value{
A character vector.
}
\description{
Format levels of confidence or probability e.g. to build variable
names.
}
\examples{
formatLevel(c(0.70, 0.95))
}
