% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonGPBayes.R
\name{poisGPBayes}
\alias{poisGPBayes}
\title{Poisson-GP Model with Bayesian Inference Results}
\usage{
poisGPBayes(data, threshold, effDuration,
            a0 = 1, b0 = 0,
            priorGP = "flatflat",
            lowerGP = c("scale" = 0, "shape" = -0.9),
            upperGP = c("scale" = Inf, "shape" = Inf),
            trace = 0)
}
\arguments{
\item{data}{Numeric vector containing the marks.}

\item{threshold}{The threshold.}

\item{effDuration}{The effective duration of the observation
period.}

\item{a0, b0}{Shape and rate for a gamma prior on the Poisson rate
\code{lambda}. Note that by choosing \code{a0 = 1} and \code{b0 =
0} we define an improper flat prior on \eqn{(0, \infty)}{(0,
Inf)}. While \code{a0} is dimensionless and compares to \eqn{1 +}
a number of observations, \code{b0} has the dimension of a
duration and compares to a duration of observation.}

\item{priorGP}{A character defining which prior is used for the GP
part of the model as in the \bold{revdbayes} package.}

\item{lowerGP, upperGP}{Bounds on the GP parameters.}

\item{trace}{Level of verbosity.}
}
\value{
An object with S3 class \code{"poisGPBayes"}.
}
\description{
Create an object representing a Poisson-GP model with Bayesian
inference result using the \strong{revdbayes} package.
}
\details{
@details
 This function mainly relies on the \strong{revdbayes} package which
 is used to produce the results of the Bayesian inference using
 MCMC. However the Maximum A Posteriori (MAP) for the parameter
 vector is computed by maximising the posterior function using a
 numerical optimisation. This allows a better comparison with the
 frequentist ML approach as used in
 \code{\link[potomax]{poisGP}}.
}
\examples{
## ========================================================================
## Use the Garonne data from Renext
## ========================================================================
fit <- poisGPBayes(data = Garonne$OTdata$Flow,
                   threshold = 2500, effDuration = 64)

## ========================================================================
## Some S3 methods: RL for Return Levels, ...
## ========================================================================
class(fit)
coef(fit)
RL(fit)
autoplot(fit)

}
\seealso{
\code{\link[potomax]{poisGP}} for a comparable object with
frequentist inference results.
}
