% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisGPBayes0.R
\name{poisGPBayes0}
\alias{poisGPBayes0}
\title{Create a Posterior for a Poisson-GP Model}
\usage{
poisGPBayes0(MCMC, threshold,
             data = NULL, effDuration,
             MAX.data = NULL, MAX.effDuration = NULL,
             OTS.data = NULL, OTS.threshold = NULL, OTS.effDuration = NULL,
             MAP = NULL,
             nOT = NA,
             a0 = 1.0, b0 = 0.0)
}
\arguments{
\item{MCMC}{An object that can be coerced into a matrix containing
the MCMC iterates. It should have the burnin period removed and be
thinned if necessary.}

\item{threshold}{the POT threshold.}

\item{data}{An optional structure containing the observations Over
the Threshold. The user should make sure that these data are
identical to those used to obtain \code{MCMC}. It can be a numeric
vector or an object inheriting from one of the two classes
\code{"Rendata"} (from \strong{Renext}) and \code{"potData"} (from
\strong{potomax}) in which case it will be coerced into the class
\code{"potData"}.}

\item{effDuration}{The observation effective duration. When
\code{data} can be coerced into \code{potData}, the provided
value for \code{effDuration} will be ignored and the value
of \code{object$OT$effDuration} will be used instead.}

\item{MAX.data, MAX.effDuration}{Optional censored data (type block
maxima or \eqn{r}-largest) as used in
\code{\link[potomax]{poisGP}}. Ignored if \code{data} inherits
from one of the two classes \code{"Rendata"} and \code{"potData"}}

\item{OTS.data, OTS.threshold, OTS.effDuration}{Optional censored
data as used in \code{\link[potomax]{poisGP}}. Ignored if
\code{data} inherits from one of the two classes \code{"Rendata"}
and \code{"potData"}}

\item{MAP}{An optional vector of Maximum A Posteriori for the
parameter vector. Should be named with names matching the
colnames of \code{MCMC}.}

\item{nOT}{An optional integer giving the number of exceedances
during the observation period. This number is needed when
\code{MCMC} does not have a column for the rate of the Poisson
process. It should be consistent with \code{MCMC}.}

\item{a0, b0}{The shape and the rate of the prior for \code{lambda}
(used only when \code{MCMC} has two columns).}
}
\value{
An object with class \code{"GEVBayes0"} inheriting from
\code{"Bayes0"}. This object can be used to produce RL plots.
}
\description{
Create an object representing a "Poor Man's" Posterior for a
Poisson-GP model, mainly using MCMC iterates.
}
\details{
The Poisson-GP model involves three parameters: one is
the rate \code{lambda} of the Poisson process describing the
exceedances over the threshold. The two other parameters are the
scale and the shape of the GP Distribution used for the excesses
over the threshold. The object MCMC can have either three columns
or two columns only. In the second case, the column corresponding
to the rate \code{lambda} is omitted; it is then assumed that
\code{lambda} is a posteriori independent of the GP parameters and
that it has a gamma posterior margin. The shape of the posterior
distribution is \eqn{a_n := a_0 + n_{\textrm{OT}}}{an = a0 + nOT}
and the rate is \eqn{b_n := b_0 + w_{\textrm{OT}}}{an = a0 + wOT},
where \eqn{w_{\textrm{OT}}}{wOT} and \eqn{w_{\textrm{OT}}}{wOT}
are the number of exceedances and the duration of the observation
period.
}
\section{Caution}{
 The user must be sure that the data provided by
using the dedicated arguments are consistent with the MCMC
iterates provided in \code{MCMC}. The 'data' dedicated arguments
are: \code{data}, \code{effDuration} or/and \code{MAX.data},
\code{MAX.effDuration} \code{OTS.data}, \code{OTS.threshold},
\code{OTS.effDuration}. Since attaching data to the object is
essentially for graphical purpose, it can be simpler not to attach
the data and to create a separated object with class
\code{"potData"}. This object can be used with its
\code{autolayer} method.
}

\examples{
require(revdbayes)
## ========================================================================
## use the 'rainfall data': daily rainfalls over 57 years
## ========================================================================
data(rainfall)
rainfall2 <- rainfall[!is.na(rainfall)]
u <- 40  ## threshold (mm)
w <- 57  ## obs duration (year)
nOT <- sum(rainfall2 > u)
prior <- set_prior(prior = "flatflat", model = "gp")
nSim <- 10000

## ========================================================================
## get a posterior for the "excess part" of the model
## ========================================================================
postGP <- rpost(n = 10000, model = "gp", prior = prior,
                data = rainfall2, thresh = u)
MCMCGP <- postGP$sim_vals
MAP <- postGP$f_mode
names(MAP) <- c("scale", "shape")

## ========================================================================
## assuming a "flat gamma prior" for 'lambda', add MCMC iterates
## for 'lambda' to those existing for GP (new matrix column)
## ========================================================================
MCMCpoisGP <- cbind(lambda = rgamma(nSim, shape = 1 + nOT, rate = 0 + w),
                    MCMCGP)

## ========================================================================
## create the object
## ========================================================================
post0 <- poisGPBayes0(MCMC = MCMCpoisGP, threshold = u, effDuration = w) 
summary(post0)
coef(post0)

## ========================================================================
## 'lambda' and the GP parameters are a posteriori independent
## ========================================================================
cov2cor(vcov(post0))

## ========================================================================
## create an object assuming 'lambda' independent of the GP param
## ========================================================================
post1 <- poisGPBayes0(MCMC = MCMCGP, threshold = u,
                      effDuration = w, nOT = nOT, MAP = MAP)

}
\seealso{
\code{\link{RL}} method to generate a data frame of
"classical" return levels (as shown on a classical RL plot),
\code{\link{predict.poisGPBayes0}} to generate a data frame of
predictive return levels (as shown on a predictive RL plot).
}
