% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEVBayes0.R
\name{predict.GEVBayes0}
\alias{predict.GEVBayes0}
\title{Predictive Quantiles or Return Levels for a GEV Model,
typically a Model for Block Maxima}
\usage{
\method{predict}{GEVBayes0}(object, newDuration = 1.0, prob,
        type = "RL",
        approx = FALSE,
        trace = 0, ...)
}
\arguments{
\item{object}{a \code{GEVBayes0} object.}

\item{newDuration}{The duration of the 'new' period for which the
maximum is to be predicted. The newduration is expressed by using
the block duration in \code{object} as unit. So \code{newDuration
= 10} means a duration of \code{10 * object$blockDuration}.}

\item{prob}{A vector of exceedance probabilities. The default
value contains such as \eqn{0.01} and \code{0.001}.}

\item{type}{The type of prediction wanted. Remind that the
\code{predict} method of the \strong{revdbayes} package allows
several types of prediction.}

\item{approx}{Logical. For the default \code{FALSE}, each value of
the tail quantile function is computed by zero-finding. For
\code{TRUE}, the quantiles are computed by using a fine of pairs
(argument, value) of the distribution function. This is likely to
be faster than \code{approx = FALSE} when \code{length(prob)} is
large.}

\item{trace}{Integer level of verbosity.}

\item{...}{Not used yet.}
}
\value{
A data frame with three columns
\item{NewDuration}{

A vector indicating the duration of the predicted period.

}
\item{Prob}{

The vector of exceedance probabilities.

}
\item{Quant}{

The vector of quantile or return levels.

}

The dataframe is given the S3 class \code{"predRL"} and it
receives several attributes such as the names of the factor
columns.
}
\description{
Compute the predictive quantiles or return levels for a GEV
Model. The quantiles are those for the maximum on a "new" period
of time \code{newDuration} years.
}
\note{
The duration \code{newDuration} is understood as given in
years and not not in block duration. Thus it sould be kept
constant when comparing Block Maxima models with different block
durations, e.g. one and two years.
}
\seealso{
\code{\link{GEVBayes0}}
}
