% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisGPBayes0.R
\name{predict.poisGPBayes0}
\alias{predict.poisGPBayes0}
\title{Predictive Quantiles or Return Levels for a EV Model
of type Poisson-GP}
\usage{
\method{predict}{poisGPBayes0}(object, newDuration = 1.0, prob,
        type = "RL",
        approx = FALSE,
        addLambda = !object$lambdaOut,
        trace = 0, ...)
}
\arguments{
\item{object}{An \code{poisGPBayes0} object, usually created by
using the eponymous creator function \code{\link{poisGPBayes0}}.}

\item{newDuration}{The duration of the 'new' period for which the
maximum is to be predicted.}

\item{prob}{A vector of exceedance probabilities. The default
value contains probabilities such as \code{0.01} and \code{0.001}.}

\item{type}{The type of prediction wanted. Remind that the
\code{predict} method of the \strong{revdbayes} package uses this
argument to allow several types of predictions: density
\code{"d"}, quantile \code{"q"}, ...}

\item{approx}{Logical. For the default \code{FALSE}, each value of
the tail quantile function is computed by zero-finding. For
\code{TRUE}, the quantiles are computed by using a fine of pairs
(argument, value) of the distribution function. This is likely to
be faster than \code{approx = FALSE} when \code{length(prob)} is
large.}

\item{addLambda}{Logical. If the description of \code{object} that
\code{'lambda'} is idependent of the GP parameters, the exact
posterior of \code{'lambda'} will by default be used to compute
the predictive distribution. But if \code{addLambda} is then
passed with ist value set to \code{TRUE}, a new colum
corresponding to \code{'lambda'} is simply added to the MCMC
iterates and the computation is carried over by ignoring the
specific independence property. In practice this seems to make
little difference.}

\item{trace}{Integer level of verbosity.}

\item{...}{Not used yet.}
}
\value{
A data frame with the following columns
\item{NewDuration}{

The duration of the "new" period on which the maximum is
predicted.

}
\item{Prob}{

A probability of exceedance.

}
\item{Quantile}{

 The return level corresponding to the probability.

}

The dataframe is given the S3 class \code{"predRL"} and it
receives several attributes such as the names of the factor
columns.
}
\description{
Prediction of an Extreme-Value model of type Poisson-GP using MCMC
iterates.
}
\details{
The rate \code{lambda} can be omitted in the MCMC iterates of
\code{object}, which will be reported via the logical flag
\code{lambdaOut} set to \code{TRUE} among the elements of the
\code{object}. Then, the parameter \code{lambda} is assumed to be
a posteriori independent of the GP parameters. In this case, the
predictive distribution can be computed by using the predictive
distribution of the number of exceedances on the new period. This
specific treatement can be by-passed by using \code{addLamba =
TRUE}; in this case a column of MCMC iterates for \code{lambda} is
added to the matrix of MCMC iterates for the GP parameters, see
the \strong{Examples} section of the \code{\link{poisGPBayes0}}
page.
}
\seealso{
\code{\link[revdbayes]{predict.evpost}} in the
\strong{revdbayes} package , and the documentation of the creator
\code{\link{poisGPBayes0}}.
}
