% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLpoisGPBayes.R
\name{RL.poisGPBayes}
\alias{RL.poisGPBayes}
\title{Return Levels and Credible Intervals for a Poisson-GP Model}
\usage{
\method{RL}{poisGPBayes}(
  object,
  period = NULL,
  level = 0.7,
  credintType = c("HPD", "eqtail"),
  smooth = missing(period),
  ...
)
}
\arguments{
\item{object}{An object with class \code{\link{poisGPBayes}} representing
the inference results for a Poisson-GP model.}

\item{period}{A vector of periods for which the return levels will
be computed.}

\item{level}{The credible level.}

\item{credintType}{The type of credible interval wanted. See
\code{\link{credInt}}.}

\item{smooth}{Logical. If \code{TRUE} the bounds of the credible
intervals are smoothed against the period.}

\item{...}{Not used yet.}
}
\value{
An object with class \code{"RL.poisGPBayes"} inheriting
from \code{"data.frame"}.
}
\description{
Compute return levels along with credible bounds for a Poisson-GP
model with Bayesian inference results.
}
\details{
Three return levels are computed, named \code{"mode"},
\code{"median"} and \code{"mean"} in relation to the posterior
distribution of each return level. The column \code{"mode"} does
not contain the value of the marginal mode of the return levels,
but rather the modal curve obtained by plugging the MAP parameter
into the quantile function. This leads to a smooth curve; the
value for a given probability or period is (usually slightly)
different from the posterior marginal mode.
}
\examples{
fit <- poisGPBayes(data = Garonne$OTdata$Flow,
                   threshold = 2500,
                   effDuration = Garonne$OTinfo$effDuration)
## Return Levels
RL(fit)
autoplot(fit)
}
\seealso{
The \code{\link[potomax]{RL}} generic function.
}
