% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLpoisGPBayes.R
\name{RL.poisGPBayes0}
\alias{RL.poisGPBayes0}
\title{Return Levels and Credible Intervals for a Poisson-GP Model 
"Poor Man's version".}
\usage{
\method{RL}{poisGPBayes0}(
  object,
  period,
  level = 0.7,
  credintType = c("HPD", "eqtail"),
  smooth = TRUE,
  ...
)
}
\arguments{
\item{object}{The \code{poisGPBayes0} object to use.}

\item{period}{See  \code{\link{RL.poisGPBayes}}.}

\item{level}{See  \code{\link{RL.poisGPBayes}}.}

\item{credintType}{The type of credible interval wanted. See
\code{\link{credInt}}.}

\item{smooth}{See \code{\link{RL.poisGPBayes}}.}

\item{...}{Other arguments to be passed to \code{\link{RL.poisGPBayes}}.}
}
\value{
An object with class \code{"RL.poisGPBayes"} inheriting
from \code{"data.frame"}.
}
\description{
Compute return levels along with credible bounds for a Poisson-GP
model with Bayesian inference results, "Poor Man's version".
}
\details{
When in the MCMC object the rate of the Poisson process is not
sampled (hence there is no column \code{"lambda"} in the matrix of
MCMC iterates) a new suitable column is added by sampling from the
marginal posterior, and the \code{RL} method for the
\code{"poisGPBayes"} class is used.
}
