% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes0.R
\name{vcov.Bayes0}
\alias{vcov.Bayes0}
\title{Covariance Matrix for a Fitted EV Model}
\usage{
\method{vcov}{Bayes0}(object, ...)
}
\arguments{
\item{object}{An object whith class \code{"Bayes0"}.}

\item{...}{Not used yet.}
}
\value{
A covariance matrix, usually of size 3.
}
\description{
Covariance matrix for the estimated parameters of an \code{Bayes0}
object.
}
\details{
The covariance is usually computed by taking the covariance of the
MCMC iterates as sored in \code{object}.
}
